% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCCEquiR.R
\name{Ineq_aggregated_level_heatmap}
\alias{Ineq_aggregated_level_heatmap}
\title{Making a heatmap with bar charts on the top and right on aggregated level data}
\usage{
Ineq_aggregated_level_heatmap(
  data,
  col,
  row,
  value,
  coln,
  rown,
  unit,
  colour = "blue"
)
}
\arguments{
\item{data}{user supplied data}

\item{col}{user defined column}

\item{row}{user defined row}

\item{value}{user defined value}

\item{coln}{user customise column name}

\item{rown}{user customise row name}

\item{unit}{user defined unit}

\item{colour}{user defined colour for the graph}
}
\value{
A graph
}
\description{
Making a heatmap with bar charts on the top and right on aggregated level data
}
